chore(tests): delete DBTrait
